Move custom exceptions to errbot.exceptions
This moves all custom exception classes to a new `errbot.exceptions`
module. This makes them easier to find and avoids the situation where
trying to catch exceptions raised by back-ends fails because of Yapsy
loading it as a new "dynamic" Python module.

Closes #418
